estudandoAPI/app_agent_test/cliente/requirements.txt
pystray
cffi
Pillow
pip
Jinja2
pywin32
six
setuptools